requirements.txt
python-dotenv
MetaTrader5